// apps/web/src/api/generated/models/OpenClawLiveAcceptanceReport.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { OpenClawLiveAcceptanceReportSummary } from './OpenClawLiveAcceptanceReportSummary';
export type OpenClawLiveAcceptanceReport = (OpenClawLiveAcceptanceReportSummary & {
    content?: string;
});