// apps/web/src/api/generated/models/OpenClawLiveAcceptanceReportSummary.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OpenClawLiveAcceptanceReportSummary = {
    fileName?: string;
    filePath?: string;
    title?: string;
    channel?: string;
    enabled?: string;
    user?: string;
    model?: string;
    target?: string;
    generatedAt?: string;
    status?: string;
    instanceId?: string;
    consoleRelayMode?: string;
    channelDeliveryMode?: string;
    messageExcerpt?: string;
};